Use size_t instead of uint for qHash
In Qt6 the hash type is now potentially using 64bit instead
of being limited to 32bit, let us follow suit.

+++ kdevplatform/language/duchain/topducontext.h
@@ -378,7 +378,7 @@ protected:
 KDEVPLATFORMLANGUAGE_EXPORT QVector<RangeInRevision> allUses(TopDUContext* context, Declaration* declaration,
                                                              bool noEmptyRanges = false);
 
-inline uint qHash(const ReferencedTopDUContext& ctx)
+inline size_t qHash(const ReferencedTopDUContext& ctx)
 {
     return ctx.hash();
 }
